Added user, admin controllers, added functions to taskservice, small fixes

+++ src/main/java/com/example/project/controller/AdminUsersController.java
@@ -2,8 +2,10 @@ package com.example.project.controller;
 
 import com.example.project.entity.User;
 import com.example.project.repository.UserRepository;
+import com.example.project.util.UserDetails;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.annotation.Secured;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 import java.util.Iterator;
@@ -14,17 +16,24 @@ import java.util.Optional;
 @Secured("hasRole('SUPER_USER')")
 public class AdminUsersController {
     @Autowired
-    UserRepository userRepo;
+    private UserRepository userRepo;
 
     @Autowired
-    PasswordEncoder encoder;
+    private PasswordEncoder encoder;
 
-    @GetMapping("/admin/users/{id}")
+    //Path const
+    private static final String path = "/admin/users/";
+
+    @GetMapping(path + "{id}")
     public String getUser(@PathVariable long id) {
-        return userRepo.findById(id).toString();
+        Optional<User> opt = userRepo.findById(id);
+        if(opt.isPresent()) {
+            return opt.get().toString();
+        }
+        return "null";
     }
 
-    @GetMapping("/admin/users/}")
+    @GetMapping(path)
     public String getAllUser() {
         Iterator<User> iter = userRepo.findAll().iterator();
         StringBuilder builder = new StringBuilder();
@@ -35,7 +44,7 @@ public class AdminUsersController {
         return builder.toString();
     }
 
-    @PostMapping("admin/users/")
+    @PostMapping(path)
     public String createUser(@ModelAttribute User user) {
         long id = userRepo.save(user).getId();
         //encode raw password
@@ -43,7 +52,7 @@ public class AdminUsersController {
         return "UserCreated with id = " + id;
     }
 
-    @PutMapping("admin/users/{id}")
+    @PutMapping(path + "{id}")
     public String updateUser(@PathVariable long id, @ModelAttribute User user) {
         Optional<User> userOpt = userRepo.findById(id);
         if (userOpt.isEmpty()) return "No user with id = " + id;
@@ -54,8 +63,11 @@ public class AdminUsersController {
         return user.toString();
     }
 
-    @DeleteMapping("admin/users/{id}")
-    public String deleteUser(@PathVariable long id) {
+    @DeleteMapping(path + "{id}")
+    public String deleteUser(@PathVariable long id, Authentication auth) {
+        UserDetails details = (UserDetails) auth.getDetails();
+        //Forbid deleting self
+        if(details.getUser().getId() == id) return "Cannot delete this user";
         userRepo.deleteById(id);
         return "User deleted";
     }
